requirements.txt
# Telegram Bot
python-telegram-bot[ext]==21.3

# EVM Web3
web3==6.20.3

# Solana
solana==0.34.3
solders==0.21.0

# HD Wallets
eth-account==0.10.0
mnemonic

# Async SQLite
aiosqlite==0.20.0

# Scheduler
APScheduler==3.10.4

# Config
python-dotenv==1.0.1
pydantic-settings==2.3.1
pydantic==2.7.4

# Cryptography
cryptography==42.0.8

# Async HTTP
httpx==0.27.0

# Logging
rich==13.7.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7

# Utilities
base58==2.1.1

# Image Generation (PnL Cards & Trade Journey)
Pillow==10.0.0

